Corrected non-entry error
This required adding an “unload” statement to the end of each macro.
The problem occurred for user forms with both text boxes and combo
boxes after a File/Save. The data wouldn’t enter until the user form
was closed. To correct this, I close the forms after each “submit”
using “Unload Me”